Scripts/.ipynb_checkpoints/requirements-checkpoint.txt
category_encoders==2.6.0
gdown==4.7.1
geopandas==1.1.1
joblib==1.3.2
mapclassify==2.4.2
matplotlib==3.8.1
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.25.2
pandas==2.3.2
Pillow==11.3.0
plotly==6.3.0
scikit_learn==1.2.2
scipy==1.16.1
seaborn==0.13.2
shap==0.42.1
streamlit==1.49.1
Unidecode==1.4.0
